Correçao do bug de adicionar, apagar, e voltar adicionar membro do grupo.

+++ src/service/GroupService.java
@@ -46,14 +46,13 @@ public class GroupService implements GroupServiceInterface {
 	@Override
 	public boolean addUserToGroup(String username, String userToAdd, String groupName) {
 		
-		System.out.println("ADDUSSERTOGROUP: "+ username + " e " + groupName + ":"+getGroupOwner(groupName));
-		
 		if (existsGroup(groupName) && getGroupOwner(groupName).equals(username)) {
+			
 			// ler ficheiro
-			String filePath = "groups/" + groupName + "/group";
-			Group group = (Group) PersistenceUtil.readObject(filePath);
+			Group group = groupDAO.getGroup(groupName);
 			System.out.println("[GroupService]" + group.getConversationId());
-
+			System.out.println("[GroupService2]" + group.getUsers());
+			
 			// utilizador nao adicionado ao grupo
 			if (!group.getUsers().contains(userToAdd)) {
 				groupDAO.addUserToGroup(userToAdd, groupName);
@@ -67,6 +66,7 @@ public class GroupService implements GroupServiceInterface {
 		}
 		// grupo nao existe
 		else {
+			
 			if(!existsGroup(groupName)) {
 				Long conversationId = groupDAO.createGroup(groupName, username);
 				groupDAO.addUserToGroup(userToAdd, groupName);
@@ -77,7 +77,6 @@ public class GroupService implements GroupServiceInterface {
 				return true;
 			}
 		}
-
 		return false;
 	}
 
@@ -86,9 +85,9 @@ public class GroupService implements GroupServiceInterface {
 		
 		// existe grupo e o utilizador eh owner
 		if (existsGroup(groupName) && getGroupOwner(groupName).equals(username)) {
+			
 			// ler ficheiro
-			String filePath = "groups/" + groupName + "/group";
-			Group group = (Group) PersistenceUtil.readObject(filePath);
+			Group group = groupDAO.getGroup(groupName);
 
 			// caso user seja o dono do grupo
 			// apagar grupo inteiro
@@ -112,11 +111,15 @@ public class GroupService implements GroupServiceInterface {
 			}
 			// apaga membro do grupo
 			else {
-				conversationService.removeConversationFromUser(userToRemove, groupName);
-				return true;
+				if(group.contains(userToRemove)) {
+					//remove a conversa do grupo das susas conversaçoes
+					conversationService.removeConversationFromUser(userToRemove, groupName);
+					//remove utilizador do grupo e persiste informação
+					groupDAO.removeUserFromGroup(group, userToRemove);
+					return true;
+				}
 			}
 		}
-
 		return false;
 	}
